Move Map.py to Outputs folder
